+++ Docs/CHANGELOG.md
@@ -18,7 +18,12 @@
 * SignalizedJunctionLeftTurn has been remade. It now has an actor flow on which the ego has to merge into, instead of a single vehicle.
 * The BackgroundActivity has been readded to the routes, which the objective of creating the sensation of traffic around the ego
 * Add waypoint reached threshold so that the precision of the actor reaching to waypoints can be adjusted based on object types.
-* **Supported OpenSCENARIO 2.0 standard.**
+* OpenSCENARIO support:
+    - Added both init and story support for `EntityAction`
+    - Added story support for `TrafficSignalControllerAction`
+    - Added both init and story support for `LateralDistanceAction`
+    - Added support for `TrafficSignalControllerCondition`
+* Supported OpenSCENARIO 2.0 standard.
 
 ### :bug: Bug Fixes
 * Fixed bug at OtherLeadingVehicle scenario causing the vehicles to move faster than intended